requirements.txt
tensorflow
pytest>=7.0.0
python-dotenv
groq
AppOpener
bs4
pillow
rich
requests
keyboard
cohere
googlesearch-python
mtranslate
pygame
edge-tts
webdriver-manager
speechrecognition
pygame
bardapi
pyperclip
pyautogui
selenium
openai
languagemodels
opencv-python-headless
pyaudio
pandas
matplotlib
cvzone
pyserial
numpy
mediapipe
googletrans==4.0.0rc1  
PyQt5
pyttsx3
nltk
torch
speedtest-cli
moviepy
beautifulsoup4
pywhatkit
google-generativeai
phonenumbers
opencage
anyio
distro